added handlebars for homepage, login, and signup and added login/signup functionality
