+++ controllers/api/userRoutes.js
@@ -1,5 +1,6 @@
 const router = require('express').Router();
 const { User } = require('../../models');
+const { Op } = require('sequelize');
 
 // CREATE new user
 router.post('/', async (req, res) => {
@@ -27,7 +28,7 @@ router.post('/login', async (req, res) => {
   try {
     const userData = await User.findOne({
       where: {
-        email: req.body.email,
+        [Op.or]: [{ email: req.body.input}, {username: req.body.input}]
       },
     });
