--- index.html
@@ -1,37 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link href='https://unpkg.com/boxicons@2.1.4/css/boxicons.min.css' rel='stylesheet'>
-    <link rel="stylesheet" href="./assets/css/style.css">
-    <title>RecipeShare</title>
-</head>
-<body>
-    <header class="">
-        <a href="" class="logo">RecipeShare</a>
-        <div>
-            <a href="./login/index.html" class="button-login">Log In</a>
-        </div>
-    </header>
-    <div class="container center">
-        <div class="main">
-            <div><h1>Recipe Catalog</h1></div>
-            <div class="search-box">
-                <input type="text" class="search-txt" placeholder="Search...">
-                <a href="#" class="search-btn"><i class='bx bx-search'></i></a>
-            </div>
-        </div>
-        <div class="button-recipe">+ Add recipe</div>
-        
-        <hr>
-
-    </div>
-
-
-
-
-
-    <script src="./assets/js/script.js"></script>
-</body>
-</html>
