--- login/index.html
@@ -1,35 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link href='https://unpkg.com/boxicons@2.1.4/css/boxicons.min.css' rel='stylesheet'>
-    <link rel="stylesheet" href="../assets/css/login.css">
-    <title>RecipeShare</title>
-</head>
-<body>
-    <div class="wrapper">
-        <form action="">
-            <h1>Log In</h1>
-            <div class="input-box">
-                <input type="text" placeholder="Username" required>
-                <i class='bx bx-user'></i>
-            </div>
-            <div class="input-box">
-                <input type="password" placeholder="Password" required>
-                <i class='bx bxs-lock-alt' ></i>
-            </div>
-            <div class="remember">
-                <label for=""><input type="checkbox">Remember me</label>
-                <a href="">Forgot password?</a>
-            </div>
-
-            <button type="submit" class="btn">Login</button>
-
-            <div class="register-link">
-                <p>Don't have an account? <a href="#">Register</a></p>
-            </div>
-        </form>
-    </div>
-</body>
-</html>
