+++ public/css/style.css
@@ -13,7 +13,7 @@ li {
 
 body {
     min-height: 100vh;
-    background: url('/assets/img/back-reg.jpg') no-repeat;
+    background: url('../img/back-reg.jpg') no-repeat;
     background-size: cover;
     background-position: center;
     border: 2px solid rgb(255, 255, 255. .2);
@@ -80,16 +80,17 @@ header {
 
 .main {
     display: flex;
+    justify-content: space-between;
     padding: 10px;
 }
 
 /* search */
 
 .search-box {
-    right: 530px;
+    /* right: 530px; */
     display: flex;
-    position: absolute;
-    height: 30px;
+    /* position: absolute; */
+    height: 50px;
     border-radius: 30px;
     padding: 10px;
     background-color: #fff;
@@ -122,11 +123,15 @@ header {
 
 /* recipe btn */
 
-.button-recipe {
-    top: 110px;
-    right: 340px;
+.buttons {
     display: flex;
-    position: absolute;
+    flex-direction: row;
+}
+.button-recipe {
+    /* top: 110px; */
+    /* right: 340px; */
+    display: flex;
+    /* position: absolute; */
 
     font-weight: 800;
     font-size: 20px;
@@ -143,6 +148,10 @@ header {
     cursor: pointer;
 }
 
+.card {
+    color: #fff;
+}
+
 
 /* line */
 
@@ -151,4 +160,97 @@ hr {
     width: 95%;
     height: 1px;
     background-color: #fff;
+}
+
+.wrapper {
+    width: 420px;
+    background: transparent;
+    color: #fff;
+    border-radius: 10px;
+    padding: 30px 40px;
+    border: 2px solid rgb(255, 255, 255. .2);
+    backdrop-filter:blur(20px);
+    box-shadow: 0 0 10px rgb(0, 0, 0. .2);
+}
+.wrapper h1 {
+    font-size: 36px;
+    text-align: center;
+}
+
+.wrapper .input-box {
+    width: 100%;
+    height: 50px;
+    margin: 30px 0;
+    position: relative;
+
+}
+
+.input-box input {
+    width: 100%;
+    height: 100%;
+    background: transparent;
+    outline: none;
+    border: 2px solid rgb(255, 255, 255. .2);
+    border-radius: 40px;
+    font-size: 16px;
+    color: #fff;
+    padding-left: 10px;
+}
+
+.input-box input::placeholder {
+    color: #fff;
+}
+
+.input-box i {
+    position: absolute;
+    right: 20px;
+    top: 50%;
+    transform: translateY(-50%);
+    font-size: 20px;
+}
+
+.wrapper .remember {
+    display: flex;
+    justify-content: space-between;
+    font-size: 14.5px;
+    margin: -15px 0 15px;
+}
+
+.remember label input {
+    accent-color: #fff;
+    margin: 3px;
+}
+
+.remember a:hover {
+    text-decoration: underline;
+}
+
+.wrapper .btn {
+    width: 100%;
+    height: 45px;
+    background-color: #fff;
+    border: none;
+    border-radius: 40px;
+    box-shadow: 0 0 10px rgb(0, 0, 0, .1);
+    cursor: pointer;
+    font-size: 16px;
+    color: #333;
+    font-weight: 600;
+}
+
+.wrapper .register-link {
+    font-size: 14px;
+    text-align: center;
+    margin-top: 20px;
+}
+
+.register-link p a{
+    color: #fff;
+    text-decoration: none;
+    font-weight: 600;
+}
+
+.register-link p a:hover {
+    text-decoration: underline;
+    color: green;
 }
