+++ views/homepage.handlebars
@@ -1,0 +1,27 @@
+
+<div class="container center">
+    <div class="main">
+        <div>
+            <h1>Recipe Catalog</h1>
+        </div>
+        <div class="buttons">
+            <div class="search-box">
+                    <input type="text" class="search-txt" placeholder="Search...">
+                    <a href="#" class="search-btn"><i class='bx bx-search'></i></a>
+            </div>
+            
+            <div class="button-recipe">
+                + Add recipe
+            </div>
+        </div>       
+    </div>
+    <hr>
+    <section class="card">
+        {{#each recipes as |recipe|}}
+            <h3><a href="/recipes/{{recipe.id}}">{{recipe.name}}</a><h3>
+            <img src="#" alt="{{recipes.imag}}">
+            <p>{{recipe.description}}</p>
+            <hr>
+        {{/each}} 
+    </section>
+</div>
