+++ views/login.handlebars
@@ -1,0 +1,24 @@
+<div class="wrapper">
+    <form class="login-form">
+        <h1>Log In</h1>
+        <div class="input-box">
+            <input id="input" type="text" placeholder="Username/Email" required>
+            <i class='bx bx-user'></i>
+        </div>
+        <div class="input-box">
+            <input id="password-login" type="password" placeholder="Password" required>
+            <i class='bx bxs-lock-alt' ></i>
+        </div>
+        <div class="remember">
+            <label for=""><input type="checkbox">Remember me</label>
+            <a href="">Forgot password?</a>
+        </div>
+
+        <button type="submit" class="btn">Login</button>
+
+        <div class="register-link">
+            <p>Don't have an account? <a href="/signup">Register</a></p>
+        </div>
+    </form>
+</div>
+<script src="/js/login.js"></script>
