+++ views/signup.handlebars
@@ -1,0 +1,19 @@
+<div class="wrapper">
+    <form class="signup-form">
+        <h1>Sign Up</h1>
+        <div class="input-box">
+            <input id="username-signup" type="text" placeholder="Username" required>
+            <i class='bx bx-user'></i>
+        </div>
+        <div class="input-box">
+            <input id="email-signup" type="text" placeholder="Email" required>
+            <i class='bx bx-user'></i>
+        </div>
+        <div class="input-box">
+            <input id="password-signup" type="password" placeholder="Password" required>
+            <i class='bx bxs-lock-alt' ></i>
+        </div>
+        <button type="submit" class="btn">Register</button>
+    </form>
+</div>
+<script src="/js/signup.js"></script>
